Fix grammar and remove some gender references from README
The readme was referencing a developer using the male pronoun "he". One could fix this by replacing "he" with "they", but since the use of a singular "they" pronoun is a bit controversial, I've instead tried to rephrase in order to completely remove gender from the text. I've also fixed some grammar mistakes and have taken the liberty to change a few words in order to make the text clearer.